Moved the isAdmin function to the authController
